services/omr-python/requirements.txt
opencv-python
numpy
fastapi
uvicorn[standard]
python-dotenv
python-multipart
